requirements.txt
PyQt6==6.7.1
numpy==1.26.4
attrs
matplotlib
mcsas3 @ git+https://github.com/BAMresearch/McSAS3.git@deef924ef0dc62dae9ad624fd82bfdd68a589280
pandas
Pint
sasmodels
scipy
PyYAML
